requirements.txt
python-dotenv
passlib[bcrypt]
fastapi
uvicorn
pydantic[email]
